requirements.txt
kafka-python
ipython
Flask
